+++ extension/src/__init__.py
@@ -1,0 +1,3 @@
+"""
+Morgen Tasks Ulauncher Extension - Source Modules
+"""